requirements.txt
Appium_Python_Client==1.3.0
